fix: moves the injection interface and symbol to separate file
