+++ src/some/some-module.provider.ts
@@ -1,17 +1,13 @@
 import { Injectable, Inject } from '@nestjs/common';
 import { SomeModuleApi } from './some-module-api.provider';
+import { SomeModuleOptions, SomeModuleOptionsSymbol } from './some.constants';
 
-export interface SomeModuleOptions {
-  apiEndpoint: string
-}
-
-export const SomeModuleOptionsSymbol = Symbol('SomeModuleOptions interface symbol for Nest DI')
 
 @Injectable()
-export class SomeModule {
+export class SomeModuleProvider {
   private readonly apiEndpoint: string
   constructor(
-    // private readonly api: SomeModuleApi, // uncomment this line for it to die horribly
+    private readonly api: SomeModuleApi, // uncomment this line for it to die horribly
     @Inject(SomeModuleOptionsSymbol) options: SomeModuleOptions,
   ) {
     console.log("Dependent provider works")
