+++ src/some/some.constants.ts
@@ -1,0 +1,5 @@
+export interface SomeModuleOptions {
+  apiEndpoint: string
+}
+
+export const SomeModuleOptionsSymbol = Symbol('SomeModuleOptions interface symbol for Nest DI')
